# Conflicts:
#	ICSCOMP1640CORE/Controllers/StaffsController.cs
#	ICSCOMP1640CORE/Views/Staffs/IdeaDetail.cshtml
#	ICSCOMP1640CORE/Views/Staffs/IdeaIndex.cshtml

+++ ICSCOMP1640CORE/Controllers/StaffsController.cs
@@ -79,6 +79,7 @@ namespace ICSCOMP1640CORE.Controllers
             var ideaInDb = _db.Ideas
                .Include(y => y.Category)
                .Include(y => y.Department)
+               .Include(y=>y.Comments)
                .Include(y => y.User)
                .Where(y => y.DepartmentId == currentUser.DepartmentId)
                .SingleOrDefault(y => y.Id == id);
@@ -172,7 +173,15 @@ namespace ICSCOMP1640CORE.Controllers
         [HttpGet]
         public IActionResult DeleteIdea(int id)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             var ideaInDb = _db.Ideas.SingleOrDefault(item => item.Id == id);
+
+            if (ideaInDb.UserId != userId)
+            {
+                _notyf.Error("You do not have permission to delete idea of other person!");
+                return RedirectToAction("IdeaIndex");
+            }
+
             _db.Ideas.Remove(ideaInDb);
             _db.SaveChanges();
             _notyf.Success("Idea is deleted successfully.");
@@ -182,14 +191,20 @@ namespace ICSCOMP1640CORE.Controllers
         [HttpGet]
         public IActionResult EditIdea(int id)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
             var model = _db.Ideas.SingleOrDefault(item => item.Id == id);
+
+            if (model.UserId != userId)
             {
-                var category = _db.Categories.ToList();
-                var categoryList = _db.Categories.Select(x => new { x.Id, x.Name }).ToList();
-                var departmentList = _db.Departments.Select(x => new { x.Id, x.Name }).ToList();
-                ViewBag.categoryList = new SelectList(categoryList, "Id", "Name");
-                ViewBag.departmentList = new SelectList(departmentList, "Id", "Name");
+                _notyf.Error("You do not have permission to edit idea of other person!");
+                return RedirectToAction("IdeaIndex");
             }
+
+            var category = _db.Categories.ToList();
+            var categoryList = _db.Categories.Select(x => new { x.Id, x.Name }).ToList();
+            var departmentList = _db.Departments.Select(x => new { x.Id, x.Name }).ToList();
+            ViewBag.categoryList = new SelectList(categoryList, "Id", "Name");
+            ViewBag.departmentList = new SelectList(departmentList, "Id", "Name");
             return View(model);
         }
 
@@ -282,5 +297,46 @@ namespace ICSCOMP1640CORE.Controllers
             _notyf.Success("Manager account is edited successfully.");
             return RedirectToAction("InforStaff");
         }
+        // COMMENTS
+        public async Task<IActionResult> ViewCommentsAsync()
+        {
+
+            var currentUser = await _userManager.GetUserAsync(User);
+            var CommentInDb = _db.Comments
+                .Include(y => y.User)
+                .ToList();
+            return View(CommentInDb);
+        }
+        [HttpGet]
+        public IActionResult CreateComments(int id)
+        {
+            Comment model = new();
+            var Comment = _db.Comments.Select(x => new { x.Id }).ToList();
+            var infoIdea = _db.Ideas.OfType<Idea>().FirstOrDefault(t => t.Id == id);
+            var Ideal = _db.Ideas.Select(x => new { x.Id }).ToList();
+            ViewBag.Comment = new SelectList(Comment, "Id", "Name");
+            ViewBag.Ideal = new SelectList(Ideal, "Id", "Name");
+            return View(model);
+        }
+        [HttpPost]
+        public async Task<IActionResult> CreateCommentsAsync(Comment comment, int id)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var infoIdea = _db.Ideas.OfType<Idea>().FirstOrDefault(t => t.Id == id);
+            var currentUser = await _userManager.GetUserAsync(User);
+            var model = new Comment();
+            {
+
+                model.IdeaId = infoIdea.Id;
+                model.UserId = userId;
+                model.Content = comment.Content;
+                model.CreatedAt = comment.CreatedAt;
+            }
+            _db.Comments.Add(model);
+            _db.SaveChanges();
+            _notyf.Success("Comment is created successfully.");
+            return RedirectToAction("IdeaIndex");
+        }
+        
     }
 }
